+++ internal/model/network.go
@@ -16,6 +16,9 @@ type Route struct {
 }
 type NetworkInterface struct {
 	Name             string   `json:"name"`
+	State            string   `json:"state,omitempty"`
+	MTU              *uint32  `json:"mtu,omitempty"`
+	MACAddress       string   `json:"mac_address,omitempty"`
 	RXBytes          uint64   `json:"rx_bytes"`
 	TXBytes          uint64   `json:"tx_bytes"`
 	RXPackets        uint64   `json:"rx_packets"`